fix: resolve CI failures for Hytte-lmqj
Extract buildDailyForecasts and related types to lib/weatherForecast.ts
so Weather.tsx only exports React components, satisfying the
react-refresh/only-export-components lint rule.
